==> notebooks/0.01_fct_enigh_download.ipynb <==
import requests
import os
import re
DOWNLOAD_PATH="data/raw/DATOS EDUCACION/ENCUESTA NACIONAL DE INGRESOS Y GASTOS EN LOS HOGARES"

os.chdir("..")

pwd

url_list=[
    #Encuesta Nacional de Ingresos y Gastos de los Hogares (ENIGH). 2016
    'https://www.inegi.org.mx/contenidos/programas/enigh/nc/2016/datosabiertos/conjunto_de_datos_enigh2016_nueva_serie_csv.zip',
    #Encuesta Nacional de Ingresos y Gastos de los Hogares (ENIGH). 2018
    'https://www.inegi.org.mx/contenidos/programas/enigh/nc/2018/datosabiertos/conjunto_de_datos_enigh_2018_ns_csv.zip',
    #Encuesta Nacional de Ingresos y Gastos de los Hogares (ENIGH) 2020
    'https://www.inegi.org.mx/contenidos/programas/enigh/nc/2020/datosabiertos/conjunto_de_datos_enigh_ns_2020_csv.zip',
    #Encuesta Nacional de Ingresos y Gastos de los Hogares (ENIGH) 2022
    'https://www.inegi.org.mx/contenidos/programas/enigh/nc/2022/datosabiertos/conjunto_de_datos_enigh_ns_2022_csv.zip'
     
    
]

def nombre_archivo(url):
    """
    url:URL de descarga del archivo que se quiere extraer el nombre

    Esta función recibe el  url de un archivo, y arroja el nombre del archivo con y sin el tipo de formato en el que esta guardado.
    """
    # Nombre con el tipo de formato incluido
    indices=[index for index in range(len(url)) if url[index]=="/"]
    last_slash=indices[-1]
    file_name=url[last_slash+1:]
    return file_name

def download_file(url,download_path):
    file_name=nombre_archivo(url)


    os.makedirs(download_path, exist_ok=True)
    response=requests.get(url)

    if response.status_code == 200:
        
            file_path=f"{download_path}/{file_name}"
            # Se guarda el contenido
            with open(file_path, 'wb') as f:
                f.write(response.content)
            #print(f'Descarga completa:{url}')
            print(f'Archivo guardado:{file_name}')
    else:
            print('Error al intentar descargar el archivo, status code:', response.status_code)


for url in url_list:
    download_file(url=url,download_path=DOWNLOAD_PATH)

import zipfile
import os

def extract_zip(zip_file_path, extract_to=None):
    """
    Extracts a ZIP file to a specified directory.
    
    Parameters:
    zip_file_path (str): The path to the ZIP file to be extracted.
    extract_to (str): The directory where the contents will be extracted.
                      If None, it extracts to the current directory.
    
    Returns:
    None
    """
    if extract_to is None:
        extract_to = os.getcwd()  # Default to current working directory

    # Ensure the extraction directory exists
    os.makedirs(extract_to, exist_ok=True)

    with zipfile.ZipFile(zip_file_path, 'r') as zip_ref:
        zip_ref.extractall(extract_to)

for file in os.listdir(DOWNLOAD_PATH):
    match=re.search(r'\d{4}',file)
    year=match.group()
    file_path=os.path.join(DOWNLOAD_PATH,file)
    extract_zip(file_path,f"{DOWNLOAD_PATH}/Extracted/{year}")
