==> notebooks/2.01_fct_ccpv_autoeda.ipynb <==
import pandas as pd
import os
from ydata_profiling import ProfileReport



os.chdir("../data/interim/CCPV/")

pwd

## 2000-2020

XXI=os.listdir("./")[4:8]
sorted_list = sorted(XXI, key=lambda x: int(x.split('.')[0]))
XXI=sorted_list
XXI

def dataframes(dataset_list):
    dataframes_list=[]
    for dataset in dataset_list:
        df=pd.read_csv(dataset)
        dataframes_list.append( (dataset, df ) )
    return dataframes_list

XXI_dataframes=dataframes(XXI)

ccpv_2000=XXI_dataframes[0][1]
ccpv_2000.drop(['Longitud', 'Latitud', 'Altitud'], axis=1, inplace=True)

ccpv_2005=XXI_dataframes[1][1]
ccpv_2005.drop(['Longitud', 'Latitud', 'Altitud'], axis=1, inplace=True)

ccpv_2010=XXI_dataframes[2][1]
ccpv_2010.drop(['Longitud', 'Latitud', 'Altitud'], axis=1, inplace=True)

ccpv_2020=XXI_dataframes[3][1]
ccpv_2020.drop(['Longitud', 'Latitud', 'Altitud'], axis=1, inplace=True)


profile_2000 = ProfileReport(ccpv_2000, title="Sonora CCPV 2000 Profile Report")
profile_2005 = ProfileReport(ccpv_2005, title="Sonora CCPV 2005 Profile Report")
profile_2010 = ProfileReport(ccpv_2010, title="Sonora CCPV 2010 Profile Report")
profile_2020 = ProfileReport(ccpv_2020, title="Sonora CCPV 2020 Profile Report")

os.makedirs("National",exist_ok=True)
profile_2000.to_file("National/ccpv_national_2000_report.html")

os.makedirs("National",exist_ok=True)
profile_2005.to_file("National/ccpv_national_2005_report.html")

os.makedirs("National",exist_ok=True)
profile_2010.to_file("National/ccpv_national_2010_report.html")

os.makedirs("National",exist_ok=True)
profile_2020.to_file("National/ccpv_national_2020_report.html")

### Sonora

sonora_2000=ccpv_2000[ccpv_2000['Clave de entidad federativa']==26]
sonora_2005=ccpv_2005[ccpv_2005['Entidad Federativa']=='Sonora']
sonora_2010=ccpv_2010[ccpv_2010['Clave de entidad federativa']==26]
sonora_2020=ccpv_2020[ccpv_2020['Clave de entidad federativa']==26]


profile_2000 = ProfileReport(sonora_2000, title="Sonora CCPV 2000 Profile Report")
profile_2005 = ProfileReport(sonora_2005, title="Sonora CCPV 2005 Profile Report")
profile_2010 = ProfileReport(sonora_2010, title="Sonora CCPV 2010 Profile Report")
profile_2020 = ProfileReport(sonora_2020, title="Sonora CCPV 2020 Profile Report")

os.makedirs("Sonora",exist_ok=True)
profile_2000.to_file("Sonora/ccpv_sonora_2000_report.html")

os.makedirs("Sonora",exist_ok=True)
profile_2005.to_file("Sonora/ccpv_sonora_2005_report.html")

os.makedirs("Sonora",exist_ok=True)
profile_2010.to_file("Sonora/ccpv_sonora_2010_report.html")

os.makedirs("Sonora",exist_ok=True)
profile_2020.to_file("Sonora/ccpv_sonora_2020_report.html")

## 1950-1995

XX=os.listdir("./")[:4]+os.listdir("./")[8:]
sorted_list = sorted(XX, key=lambda x: int(x.split('.')[0]))
XX=sorted_list
XX
