==> notebooks/2.01_fct_enigh_autoeda.ipynb <==
import requests
import os
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import math
import numpy as np
from ydata_profiling import ProfileReport
INTERIM_PATH="data/interim/ENIGH"

os.chdir(f'../{INTERIM_PATH}')

df_2022=pd.read_csv("2022.csv")
df_2020=pd.read_csv("2020.csv")
df_2018=pd.read_csv("2018.csv")
df_2016=pd.read_csv("2016.csv")

def in_sonora(df):
    in_sonora= (df['folioviv']>=2600000000)&(df['folioviv']<2700000000)
    return in_sonora

## ENIGH 2022 Dataset

df=df_2022[[
    'folioviv', 'foliohog', 'numren', 'parentesco', 'sexo',
    'edad', 'madre_hog', 'madre_id', 'padre_hog', 'padre_id',
    'alfabetism',
    'asis_esc', 'nivel', 'grado', 'tipoesc', 'tiene_b', 'otorg_b',
    'forma_b', 'tiene_c', 'otorg_c', 'forma_c', 'nivelaprob', 'gradoaprob',
    'antec_esc', 'residencia', 'edo_conyug', 'pareja_hog', 'conyuge_id',
    'año_nacimiento', 'generacion', 'index', 'gasto_tri', 'ing_tri',
       'clave', 'numero_de_ingresos_hogar', 'ing_tri_hogar_total',
       'ing_tri_individual_total','generacion_conyuge', 'delta'

]]
profile = ProfileReport(df, title="ENIGH 2022")

os.makedirs("National",exist_ok=True)
profile.to_file("National/National_2022_report.html")

### Sonora 


profile = ProfileReport(df[in_sonora(df)], title="ENIGH 2022")
os.makedirs("Sonora",exist_ok=True)
profile.to_file("Sonora/Sonora_2022_report.html")

## ENIGH 2020 dataset

df=df_2020[[
    'folioviv', 'foliohog', 'numren', 'parentesco', 'sexo',
    'edad', 'madre_hog', 'madre_id', 'padre_hog', 'padre_id',
    'alfabetism',
    'asis_esc', 'nivel', 'grado', 'tipoesc', 'tiene_b', 'otorg_b',
    'forma_b', 'tiene_c', 'otorg_c', 'forma_c', 'nivelaprob', 'gradoaprob',
    'antec_esc', 'residencia', 'edo_conyug', 'pareja_hog', 'conyuge_id',
    'año_nacimiento', 'generacion', 'index', 'gasto_tri', 'ing_tri',
       'clave', 'numero_de_ingresos_hogar', 'ing_tri_hogar_total',
       'ing_tri_individual_total','generacion_conyuge', 'delta'

]]
os.makedirs("National",exist_ok=True)
profile = ProfileReport(df, title="ENIGH 2020")
profile.to_file("National/National_2020_report.html")

### Sonora

profile = ProfileReport(df[in_sonora(df)], title="ENIGH 2020")
os.makedirs("Sonora",exist_ok=True)
profile.to_file("Sonora/Sonora_2020_report.html")

## ENIGH 2018 dataset

df=df_2018[[
    'folioviv', 'foliohog', 'numren', 'parentesco', 'sexo',
    'edad', 'madre_hog', 'madre_id', 'padre_hog', 'padre_id',
    'alfabetism',
    'asis_esc', 'nivel', 'grado', 'tipoesc', 'tiene_b', 'otorg_b',
    'forma_b', 'tiene_c', 'otorg_c', 'forma_c', 'nivelaprob', 'gradoaprob',
    'antec_esc', 'residencia', 'edo_conyug', 'pareja_hog', 'conyuge_id',
    'año_nacimiento', 'generacion', 'index', 'gasto_tri', 'ing_tri',
       'clave', 'numero_de_ingresos_hogar', 'ing_tri_hogar_total',
       'ing_tri_individual_total','generacion_conyuge', 'delta'

]]
os.makedirs("National",exist_ok=True)
profile = ProfileReport(df, title="ENIGH 2018")
profile.to_file("National/National_2018_report.html")

### Sonora

profile = ProfileReport(df[in_sonora(df)], title="ENIGH 2018")
os.makedirs("Sonora",exist_ok=True)
profile.to_file("Sonora/Sonora_2018_report.html")

## ENIGH 2016 dataset

df=df_2016[[
    'folioviv', 'foliohog', 'numren', 'parentesco', 'sexo',
    'edad', 'madre_hog', 'madre_id', 'padre_hog', 'padre_id',
    'alfabetism',
    'asis_esc', 'nivel', 'grado', 'tipoesc', 'tiene_b', 'otorg_b',
    'forma_b', 'tiene_c', 'otorg_c', 'forma_c', 'nivelaprob', 'gradoaprob',
    'antec_esc', 'residencia', 'edo_conyug', 'pareja_hog', 'conyuge_id',
    'año_nacimiento', 'generacion', 'index', 'gasto_tri', 'ing_tri',
       'clave', 'numero_de_ingresos_hogar', 'ing_tri_hogar_total',
       'ing_tri_individual_total','generacion_conyuge', 'delta'

]]
os.makedirs("National",exist_ok=True)
profile = ProfileReport(df, title="ENIGH 2016")
profile.to_file("National/National_2016_report.html")

### Sonora

profile = ProfileReport(df[in_sonora(df)], title="ENIGH 2016")
os.makedirs("Sonora",exist_ok=True)
profile.to_file("Sonora/Sonora_2016_report.html")
