==> notebooks/2.03_fct_enigh_outliers.ipynb <==
import requests
import os
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import math
import numpy as np

INTERIM_PATH="data/interim/ENIGH"
RAW_PATH="data/raw/DATOS EDUCACION/ENCUESTA NACIONAL DE INGRESOS Y GASTOS EN LOS HOGARES/Extracted"


os.chdir('..')

data_path=f"{INTERIM_PATH}/2022.csv"
df_2022=pd.read_csv(data_path)


df_2022.info()

df_2022.columns[:100]

df_2022.columns[100:200]

df_2022=df_2022[ [
                 # Integer
                'folioviv', 'foliohog', 'numren', 'parentesco', 'edad',
                'madre_hog', 'madre_id', 'padre_hog','num_trabaj','pareja_hog','conyuge_id',
                
                #Float
                'gasto_tri', 'ing_tri','peso',

                #Categorical
                'alfabetism','generacion','entidad','asis_esc','grado','nivel','tipoesc','gradoaprob','clave',
                'sexo','nivelaprob','edo_conyug',
                
                #String


                
                
                 
                'antec_esc'
                
                ] ]

df_2022.info()

df_2022.antec_esc.unique()

df_2022=df_2022[df_2022['alfabetism']!=' ']
df_2022['alfabetism']=df_2022['alfabetism'].astype('int')

df=df_2022[ [
                 # Integer
                'folioviv', 'foliohog', 'numren', 'edad',
                'madre_hog', 'padre_hog','num_trabaj','pareja_hog',
                
                #Float
                'gasto_tri', 'ing_tri','peso',

                #Categorical
                'alfabetism','generacion','entidad','grado','nivel','tipoesc','gradoaprob','clave',
                'sexo','nivelaprob','edo_conyug','asis_esc',
                
                #String


                
                
                 
                'antec_esc',   
                
                ] ]

df.info()

## num_trabaj

df.num_trabaj.unique()

df[df['num_trabaj']==' '].clave.unique()

From what I can tell, ' ' in the 'num_trabaj' column means the individual is not currently working at the time of the interview. We'll replace each of these with 'No job' in the categorical column that will replace it. 

## pareja_hog

df.pareja_hog.unique()

df[df['pareja_hog']==' '].edo_conyug.unique()

df[df['edo_conyug']==' '].edad.unique()

From what I can tell, ' ' in the 'pareja_hog' column means the individual is not currently living with their couple at the time of the interview. We'll replace each of these with 'No couple' in the categorical column that will replace it. 

## grado

column_name='grado'

df[column_name].unique()

df[df[column_name]==' '].nivelaprob.unique()

df[df[column_name]==' '].nivel.unique()

df[(df[column_name]==' ') & (df['nivel']=='06')]


df[(df[column_name]==' ') & (df['nivel']=='04')]

df[df[column_name]==' '].asis_esc.unique()

df[(df[column_name]==' ') &( df['asis_esc']=='1')]

In this case, by looking at the ocurrences of ' ' in the column 'grado', it would appear as though it meant that the subject is not currently assisting school.  

## nivel

column_name='nivel'

df[column_name].unique()

df[(df[column_name]==' ')].asis_esc.unique()

df[(df[column_name]==' ') &( df['asis_esc']=='1')]

Same case as the column 'grado' 

## tipoesc

column_name='tipoesc'

df[column_name].unique()

df[(df[column_name]==' ')].asis_esc.unique()

Same case as column 'nivel'

## gradoaprob


column_name='gradoaprob'


df[column_name].unique()


Just need to format the column correctly

## nivelaprob

column_name='nivelaprob'


df[column_name].unique()


df[(df[column_name]==' ')].asis_esc.unique()

Just need to format the column correctly

## edo_conyug

column_name='edo_conyug'


df[column_name].unique()


df[(df[column_name]==' ')].edad.unique()

The ocurrences of ' ' happen for individuals with less than 11 years old.

## antec_esc

column_name='antec_esc'


df[column_name].unique()


df[(df[column_name]==' ')]

# PCA

df_2022=df_2022[df_2022['alfabetism']!=' ']
df_2022['alfabetism']=df_2022['alfabetism'].astype('int')

df=df_2022[ [
                 # Integer
                'folioviv', 'foliohog', 'numren', 'edad',
                'madre_hog', 'padre_hog',
                
                #Float
                'gasto_tri', 'ing_tri','peso',

                #Categorical
                'alfabetism','generacion','entidad',
                'sexo',
                
                #String


                
                
                   
                
                ] ]

df['generacion']=df['generacion'].astype('category')
df['alfabetism']=df['alfabetism'].astype('category')
df['entidad']=df['entidad'].astype('category')
df['sexo']=df['sexo'].astype('category')
 
               

df.info()

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.decomposition import PCA

n_components=4
categorical_columns = df.select_dtypes(include=['category']).columns
numerical_columns = df.select_dtypes(include=['number']).columns

# Define the preprocessor
preprocessor = ColumnTransformer(
    transformers=[
        ('num', StandardScaler(), numerical_columns),
        ('cat', OneHotEncoder(), categorical_columns)
    ]
)



pipe = Pipeline([
    ('preprocessor', preprocessor),
    ('pca', PCA(n_components=n_components))
])


pipe.fit(df)


pca = pipe.named_steps['pca']

encoded_feature_names = (
    numerical_columns.tolist() +
    pipe.named_steps['preprocessor'].named_transformers_['cat']
    .get_feature_names_out(categorical_columns).tolist()
)

pca_df = pd.DataFrame(
    pca.components_.T,
    columns=[f'PC {x}' for x in range(1, n_components + 1)],
    index=encoded_feature_names
)



pca_df

import numpy as np
import plotly.express as px

print("Varianza por componente: ", pca.explained_variance_)
print("Porcentaje de varianza explicada por componente: ", 
      pca.explained_variance_ratio_)

df_ev = pd.DataFrame({
    "Varianza explicada": np.cumsum(pca.explained_variance_ratio_),
    "Componente": [f'PC {x}' for x in range(1,n_components+1)]
})
fig = px.line(df_ev, x='Componente', y='Varianza explicada', markers=True,
              title="Varianza explicada por componente")
fig.show()

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.decomposition import PCA

n_components=6
categorical_columns = df.select_dtypes(include=['category']).columns
numerical_columns = df.select_dtypes(include=['number']).columns

# Define the preprocessor
preprocessor = ColumnTransformer(
    transformers=[
        ('num', StandardScaler(), numerical_columns),
        ('cat', OneHotEncoder(), categorical_columns)
    ]
)



pipe = Pipeline([
    ('preprocessor', preprocessor),
    ('pca', PCA(n_components=n_components))
])


pipe.fit(df)


pca = pipe.named_steps['pca']

encoded_feature_names = (
    numerical_columns.tolist() +
    pipe.named_steps['preprocessor'].named_transformers_['cat']
    .get_feature_names_out(categorical_columns).tolist()
)

pca_df = pd.DataFrame(
    pca.components_.T,
    columns=[f'PC {x}' for x in range(1, n_components + 1)],
    index=encoded_feature_names
)



import numpy as np
import plotly.express as px

print("Varianza por componente: ", pca.explained_variance_)
print("Porcentaje de varianza explicada por componente: ", 
      pca.explained_variance_ratio_)

df_ev = pd.DataFrame({
    "Varianza explicada": np.cumsum(pca.explained_variance_ratio_),
    "Componente": [f'PC {x}' for x in range(1,n_components+1)]
})
fig = px.line(df_ev, x='Componente', y='Varianza explicada', markers=True,
              title="Varianza explicada por componente")
fig.show()

# LOF 

from sklearn.neighbors import LocalOutlierFactor

clf = LocalOutlierFactor(n_neighbors=20, contamination=0.1)
y_pred = clf.fit_predict(df)
#n_errors = (y_pred != ground_truth).sum()
X_scores = clf.negative_outlier_factor_

from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.decomposition import PCA

# Identify columns
categorical_columns = df.select_dtypes(include=['object', 'category']).columns
numerical_columns = df.select_dtypes(include=['number']).columns

# Define a preprocessor
preprocessor = ColumnTransformer(
    transformers=[
        ('num', StandardScaler(), numerical_columns),
        ('cat', OneHotEncoder(), categorical_columns)
    ]
)

# Create a pipeline with LOF
pipeline = Pipeline([
    ('preprocessor', preprocessor),
    ('lof', LocalOutlierFactor(n_neighbors=20))
])

# Fit the pipeline
pipeline.fit(df)



# Fit and transform the data using the preprocessor
transformed_data = pipeline.named_steps['preprocessor'].fit_transform(df)

# Get feature names from the preprocessor (use get_feature_names_out for OneHotEncoder)
encoded_feature_names = (
    numerical_columns.tolist() +
    pipeline.named_steps['preprocessor'].named_transformers_['cat']
    .get_feature_names_out(categorical_columns).tolist()
)

# Create a DataFrame with the transformed data
transformed_df = pd.DataFrame(transformed_data, columns=encoded_feature_names)

# Display the transformed DataFrame
print(transformed_df.head())
