==> notebooks/2.04_kgm_Formated_Data.ipynb <==
# Poner en orden los datos y hacer un EDA rapido

Importamos las librerias

import os
import pandas as pd
from datetime import datetime
import matplotlib.pyplot as plt
import seaborn as sns
import ipywidgets as widgets
from ipywidgets import interact
import geopandas as gpd

df_f = pd.read_csv('ENIGH.csv')
mapa = os.path.join('..', '..', 'ingCaracteristicasProyecto_', 'data', 'raw', 'MAPA', '2023_1_00_ENT.shp')

'''Cargar el CSV y el archivo .shp'''
datos_mapa = gpd.read_file(mapa, encoding='latin-1')

df_f['Entidad'] = df_f['entidad'].astype(int)
datos_mapa['CVE_ENT'] = datos_mapa['CVE_ENT'].astype(int)

    # Fusionar los DataFrames en la columna 'Entidad' y 'CVE_ENT'
df_f = pd.merge(df_f, datos_mapa, left_on='Entidad', right_on='CVE_ENT', how='left')

    # Convertir a GeoDataFrame y renombrar columnas
df_f = gpd.GeoDataFrame(df_f, geometry='geometry')
df_f = df_f.drop(columns=['CVE_ENT', 'CVEGEO', 'geometry', 'entidad'])
df_f= df_f.rename(columns={'NOMGEO': 'Entidad', 'Entidad': 'C_Entidad', 'folioviv':'Folio', 'edad':'Edad',	'sexo':'Sexo', 'year':'Fecha', 'alfabetism':'Alfabetismo',	'generacion':'Generacion'})
df_f = df_f[['Folio', 'Edad', 'Sexo', 'C_Entidad', 'Entidad', 'Fecha', 'Alfabetismo', 'Generacion']]
# Cambiar los tipos de datos
df_f['Folio'] = df_f['Folio'].astype('object')
df_f['Edad'] = df_f['Edad'].astype('int64')    # Asegurar que es entero
df_f['Sexo'] = df_f['Sexo'].astype('category')  # Convertir a categoría
df_f['C_Entidad'] = df_f['C_Entidad'].astype('category')  # Convertir a categoría
df_f['Alfabetismo'] = df_f['Alfabetismo'].astype('category')
df_f['Fecha'] = pd.to_datetime(df_f['Fecha'], format='%Y')


df_f.to_parquet('enigh_processed.parquet')

A continuacion `df_1` son los datos de salud mental, alcoholismo y tabaquismo, mientras que `df_2` son los resultados de la encuesta del **enigh**.

df_1 = pd.read_parquet('Ensanut-data-processed.parquet')
df_2 = pd.read_parquet('enigh_processed.parquet')

df_1['Sexo'] = df_1['Sexo'].astype('category')
df_1['C_Entidad'] = df_1['C_Entidad'].astype('category')
df_1['Fecha'] = pd.to_datetime(df_1['Fecha'], errors='coerce')
df_1['Año'] = df_1['Fecha'].dt.year
df_1 = df_1.dropna(subset=['Edad']).reset_index(drop=True)
df_1['Edad'] = df_1['Edad'].astype('int')


df_2['Sexo'] = df_2['Sexo'].astype('category')
df_2['C_Entidad'] = df_2['C_Entidad'].astype('category')
df_2['Fecha'] = pd.to_datetime(df_2['Fecha'], format='%Y')
df_2['Año'] = df_2['Fecha'].dt.year
df_2['Generacion'] = df_2['Generacion'].astype('category')

df_1.head()

df_2['Año'].unique()

df_2['Generacion'].unique()

def asignar_generacion(edad):
    if edad < 10:
        return "Alpha"
    elif 10 <= edad < 20:
        return "Z"
    elif 20 <= edad < 36:
        return "Milenial"
    elif 36 <= edad < 52:
        return "X"
    elif 52 <= edad < 71:
        return "Baby Boomer"
    elif 71 <= edad < 92:
        return "Silent"
    elif edad >= 92:
        return "Greatest"
    else:
        return "Unknown"

df_1['Generacion'] = df_1['Edad'].apply(asignar_generacion)
df_1.head()

Vamos a dejar solo en los que el anio sea 2016, 2020 o 2022

df_1 = df_1[df_1['Año'].isin([2018, 2021, 2022])].reset_index(drop=True)
df_2 = df_2[df_2['Año'].isin([2018, 2020, 2022])].reset_index(drop=True)
df_1.head()

df_1['Año'] = df_1['Año'].replace({2021: 2020})

df_1.loc[df_1["Año"] == 2020, "Fecha"] = "2020-07-31"

df_1['Fecha'] = pd.to_datetime(df_1['Fecha'], errors='coerce')

print(df_1['Año'].unique())


df_1.to_parquet('ensanut_processed.parquet')
df_2.to_parquet('enigh_processed.parquet')
